Fix printing on Foundation site skins
Printing an entry+comments page on the new site skins pretty much just EXPLODES.
Shoutout to anaraine and AlexSeanchai in the Discord for raising the issue.

First set of problems is in Foundation's print styles, which "helpfully" add the
URL after every link (including tiny chiclet icon links that carry no meaning in
print). So let's turn those off and add our own reduced set, because flipping
font to black and expanding abbreviations is still a pretty good idea.

Second set of problems involves the positioning of account links and the user's
icon in the header, but those are straightforward fixes too.

Third set of problems is specific to Tropo, which cuts off everything after the
first page. 😱 Addressed in dreamwidth/dw-nonfree#142.

After that we get into the innate problems that already existed on the old skins
-- showing threading through indentation alone makes a truly impossible mess
when printing deep threads, and that can't be fixed without a total rethink of
the threading styles. But at least this commit (plus nonfree 142) gets us back
to status quo ante.