Foundation modals: clear some space around the close button
In the last big foundation update, something changed with how things are spaced
in reveal modals, which made it possible for the edges of elements (like the
top border of a fieldset) to sometimes scoot underneath the close button. That
looks a little confusing, so let's make the close button obscure the couple of
pixels behind it by grabbing the same background color as the modal itself.