Fix icon browser on create entries page (#2836)
* Un-break the icon browser in beta create entries page

Somehow I missed this bug in my last icon browser PR.

Visible symptoms: "order by keyword" doesn't work when used on the create
entries page, and some of the styling/alignment is a little wrong there too.

Underlying problem:

- The whole create entries page is wrapped in one bigass `<form>` element.
- It's illegal to nest a form inside a form.
- I used forms as container elements for the re-jiggered icon browser options toggles.
- The icon browser modal markup gets spawned _inside_ the icons module, which is
  still inside the bigass top-level form element.
- Kaboom.

I assume this also broke some other stuff on that page, but who cares because
let's just not do that; those nested form elements aren't needed anyway. (I used
them initially because I thought I'd be using `HTMLFormElement.elements` to get
the inputs, but I didn't end up doing that.)

* Create entries: Use normal fieldset/legend relationship in custom access modal

This was making the top-border on Foundation's fieldset styling look wrong.

* Create entries: Don't mess with fieldset borders in modals

This was causing the icon browser to look inconsistent between create entries
and site-skin comment pages. After following it back, it looks like this was
just to accommodate the weird markup in the custom access groups modal.

* Foundation modals: clear some space around the close button

In the last big foundation update, something changed with how things are spaced
in reveal modals, which made it possible for the edges of elements (like the
top border of a fieldset) to sometimes scoot underneath the close button. That
looks a little confusing, so let's make the close button obscure the couple of
pixels behind it by grabbing the same background color as the modal itself.

* Create entries: Fix button alignment when using custom access groups

The fancy selects squirted a few pixels out of alignment whenever the "edit"
button appeared; this stops that.