Two minor reply form tweaks (#2831)
* Comment preview: use same reply form font as other site skin pages

The rules that affect entries and comments weren't affecting the parent wrapper
for the preview page. That was mostly fine, because the preview itself and the
parent post both use fake wrappers to make sure they get the right styles. But
the talkform appears outside those, as its own element, so it was missing the
rules. Thus, the textarea was at the mercy of whatever default or inherited
styles applied, probably browser-dependent.

* Reply form: Allow a little more max-width on desktop

On site-skin comment pages, the 16px font for textareas just makes the max
starting width feel too cramped.