Restore selected icon when editing an entry
The form element template helpers automatically populate the `selected`
attribute of a select based on the `formdata` argument of the parent template
that called the helper, if it's present and it contains a key that matches the
`name` of the select. And the entry page took advantage of that behavior for the
icons menu. Passing `selected` breaks that if the `selected` value is undef.

As of this commit, the icon preview is still bugged when editing an entry -- the
select is set to the right icon and the preview gets updated if you change it,
but it starts out on the default icon no matter what.