Add class="table" to a bunch of tables in views/ (but not all of em)
- If it has any deprecated pre-CSS styling attributes like cellpadding or
cellspacing, it obviously wasn't written with the expectation of being inserted
into a modern website with global table styles, so it's in the same bucket as
everyone's copy-pasted character sheet markup and does not want any schmancy
default table styles.

- If it has a style attribute that affects the interior content at all, it also
probably predates the foundation site skins and I don't want to bother with it.
But if it's just messing with the margins, that's probably fine.