Plack testing
This makes the minimal changes necessary to enable supporting Plack as a
runtime environment instead of Apache.

The idea is to make running the code much, much simpler since we can
have a pure Perl web server and never need to deal with system modules
and the increasingly difficult to deal with Apache/mod_perl stuff.

However, Plack is a pluggable runtime so it still does support mod_perl
so we can do that for efficiency in production if we want. Or, we can
start to look at something like Starman.