Remove credentials
It looks like Paws does this automatically, using the standard method of
setting AWS configs -- the environment variables, the .aws files in
homedir, or via instance profiles.

I'm doing this here because I'm trying to get ECS credentials to work,
and instead of using yet a third credential library here, I'm going to
try this the default way.