Fix email sending duplication logic
The 'don't duplicate emails' logic was too conservative. If we hit a
transient error later sending the email (which happens all the time),
then we would never actually retry because we set the key too early.

This fixes it to move setting the MemCache bit until the very end of the
email flow. This means there's a tiny risk of us duplicating an email,
but this does make it more of an at-least-once system (rather than an
at-most-once system, which is not great for email.)