Simple Captcha
This adds a very simple site-wide captcha (but only on things that go
thru Controller...)

The goal here is to make it so we can replace the external captcha
system that we have today with something that is more controllable and
user friendly, as hCaptcha is pretty accessible.

Next up, I will try to standardize our captchas so that we have only the
one system for doing it, instead of doing it in N different places.