Add simple auto-blocking of fraudsters
Bots and scrapers that just get captcha redirects over and over but
never attempt to solve will get blocked. This should be fairly loose so
as not to cause user impact, but will require tuning.

This also adds a simple memcache based tempban system which doesn't
persist to the database.