bug in LatestNews widget on logged-in homepage
Obscure, but I just ran into this in testing: if you
set $LJ::NEWS_JOURNAL to a valid community but it
doesn't have any entries yet, the widget will bomb
out with a missing anum error.

This adds a validity check to the entry object before
attempting to do anything with it.