allow for=self paid time buyers to future date purchases
We were hiding all of the "gift" options on this version of
the form, but there are legit reasons for someone to want to
future-date a purchase for themselves. You could work around
it by visiting the for=gift version of the page and entering
your own username, but that's pointlessly annoying.