limit directory search by location results to visible users
I took my cue from ContactInfo.pm, which mentions a desire
to avoid using load_userids on a large data set for filtering
purposes. Instead this expresses the visibility requirement as
an additional WHERE clause in the SELECT statement, acting only
on users that have rows in both the user and userprop tables.

Obviously my ability to test this locally is extremely limited,
but this should fix the problem with location results being
sparsely populated due to a large number of suspended SEO accounts.