[#2921] fix table header background color on Foundation pages
I found where the table header link colors were specified
and dropped in a rule for a background color so that the
links would be legible again.

It was very obvious on the admin/statushistory page, but
this should affect any Foundation pages with similarly
defined table header elements.

I don't know if this is the best solution, but it seems
like an improvement. I did check legibility in all the
different site schemes.