Update comments to match license stated, add link
Although the license stated in the text is CC-by-SA-2.5, the comments in the header incorrectly stated contradictory information apparently left over from use of that header in a code file. This pull request updaes the comments to be consistent with the license stated, and adds a link to the license as required for its use. I would like to fix these issues before making a copy to use on our own site with appropriate alterations and credit as per the license.