Allow logging out multiple sessions/all sessions (#3016)
* Fix loginlog with long ua's

This trims useragents so they fit into the VARCHAR(100) that is the db
column. This also fixes so that we will log errors when we fail to
record logins, so this won't show up silently in production.

* Add multi-select and logout all

This adds the ability to log out multiple sessions at the same time as
well as to log out all sessions in one click, including the one you're
actively using.