more generic fix for legacy country codes (#3031)
Years ago, we changed our source of country code data to use a
standard CPAN module.  This has largely worked well except for a
couple of odd corner cases where the previous country code stored in
user data didn't match up with the new source.  The obvious one was
the change from UK to GB, so we kludged in recognition for both codes,
and then deleted the older one anywhere it showed up as a duplicate in
a drop down menu.

Subsequently it was noted that there was a legacy code for Scotland
(LJSC) which needed the same treatment. Instead of continuing to
expand the kludge, this adds a second class method, load_legacy, to
use when requesting the version of the data that includes the older
codes. That way we can go back to calling the basic load method
without having to then delete the extra codes in the caller. And if we
stumble across any other missing codes, they'll be maintained in one
place.

Fixes #2197.