new GoogleAnalytics4 user setting
This is the bare minimum of code needed to allow users to tell us
their IDs for use with the new version of Google Analytics.
It doesn't do anything with that information yet.

* new module cgi-bin/DW/Setting/GoogleAnalytics4.pm, which is more or less
  a clone of cgi-bin/DW/Setting/GoogleAnalytics.pm with tweaks

* both settings modules are retained to allow for a transition period;
  the labels on /manage/settings are "Google Analytics ID" and
  "Google Analytics ID (v4)"

* new userprop ga4_analytics

* new user method ga4_analytics

I couldn't find a format guideline for the ID values. The setting module
enforces a format of "G-" followed by at least 10 alphanumeric characters,
saved as uppercase. If that turns out to be wrong, we can tweak it later.