fix failing t/media-security.t (#3127)
A line I added in #3100 was failing in testing because there was no
web request object defined in that context. This was the error:

Can't call method "headers_in" on an undefined value at
dw/cgi-bin/DW/Media/Base.pm line 140.

This adds an undefined value check to make sure everything proceeds smoothly.