De-jQuery shop
This converts jquery.shop.js to just be plain js and removes jquery from
the entire shop pages.

This breaks the dropdowns on the menus which will have to do for now and
we can figure out how to do a non-jquery version of this I guess. I also
made it slightly forward looking so we can start removing jquery from
more foundation pages if we want.