Simplify email system
This causes the old TheSchwartz worker to simply enqueue an SQS job in
the new task system, so we don't have any duplicate mail sending logic.
Additionally, this standardizes on Net::SMTP which is the modern module
now that it supports TLS. Also fix up the configs so this doesn't
reference SES anymore since it's not specific to that.