Remove Procnotify references and related code
- Deleted or updated all code referencing LJ::Procnotify
- Cleaned up bin scripts and modules to remove Procnotify usage
- Updated Sysban, ljlib, and user rename logic accordingly

This is reasonable because our Apache processes automatically exit
after a certain period of time (100 requests or memory limit) which caps
the length of time that they're possibly out of date, and we use memcache
for most lookups which doesn't require synchronization.

So the "risk" here is pretty minimal IMO and we aren't really using this
system going forward.